% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_language_codes.R
\name{extract_language_codes}
\alias{extract_language_codes}
\title{Extract buko Language Code from Inspector Course Names}
\usage{
extract_language_codes(inspector_data)
}
\arguments{
\item{inspector_data}{a data frame including a field course_name}
}
\description{
This function uses a regular expression to find a 4-6 digit code in the course_name field of a data frame
}
\examples{
extract_language_codes()
}
\keyword{buko}
\keyword{inspector}
