% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_project_ids.R
\name{extract_project_ids}
\alias{extract_project_ids}
\title{Extract buko project ID from Inspector Course Names}
\usage{
extract_project_ids(inspector_data)
}
\arguments{
\item{inspector_data}{a data frame including a field course_name}
}
\description{
This function uses a regular expression to find a 4-6 digit code in the course_name field of a data frame
}
\examples{
extract_project_ids()
}
\keyword{buko}
\keyword{inspector}
